requirements.txt
Flask==3.0.0
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
fpdf==1.7.2
plotly==5.18.0
python-dotenv==1.0.0

mysql-connector-python

